# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinjie Fullstack Base Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Sign in a browser user",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserAuthSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Rotate the browser refresh token",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "pinjie_web_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Web Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Sign out the current browser session",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "pinjie_web_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Web Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Sign in an administrator",
        "operationId": "login_api_v1_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminAuthSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/refresh": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Rotate administrator refresh token",
        "operationId": "refresh_api_v1_admin_auth_refresh_post",
        "parameters": [
          {
            "name": "pinjie_admin_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Admin Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Sign out the current administrator session",
        "operationId": "logout_api_v1_admin_auth_logout_post",
        "parameters": [
          {
            "name": "pinjie_admin_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Admin Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Get the current administrator",
        "operationId": "get_me_api_v1_admin_auth_me_get",
        "parameters": [
          {
            "name": "pinjie_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/password": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Change administrator password",
        "operationId": "change_password_api_v1_admin_auth_password_post",
        "parameters": [
          {
            "name": "pinjie_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/confirm": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Confirm a sensitive administrator action",
        "operationId": "confirm_api_v1_admin_auth_confirm_post",
        "parameters": [
          {
            "name": "pinjie_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminConfirmIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminConfirmOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminAuthSessionOut": {
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/AdminRead"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "principal",
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "AdminAuthSessionOut"
      },
      "AdminConfirmIn": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password"
          },
          "action": {
            "$ref": "#/components/schemas/ConfirmationAction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "action"
        ],
        "title": "AdminConfirmIn"
      },
      "AdminConfirmOut": {
        "properties": {
          "confirmation_token": {
            "type": "string",
            "title": "Confirmation Token"
          },
          "action": {
            "$ref": "#/components/schemas/ConfirmationAction"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmation_token",
          "action",
          "expires_at"
        ],
        "title": "AdminConfirmOut"
      },
...
      "AdminLoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "AdminLoginIn"
      },
      "AdminRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            },
            "type": "array",
            "title": "Roles"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "is_active",
          "is_superuser",
          "roles",
          "permissions",
          "created_at",
          "updated_at"
        ],
        "title": "AdminRead"
      },
...
      "ConfirmationAction": {
        "type": "string",
        "enum": [
          "users:disable",
          "users:credentials:reset",
          "users:sessions:revoke",
          "admins:create",
          "admins:superuser:change",
          "admins:status:change",
          "admins:credentials:reset",
          "admins:roles:assign",
          "admins:sessions:revoke",
          "roles:delete",
          "roles:permissions:assign"
        ],
        "title": "ConfirmationAction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChangeIn": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeIn"
      },
...
      "RefreshSessionOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "RefreshSessionOut"
      },
...
      "ResponseModel_AdminAuthSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/AdminAuthSessionOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminAuthSessionOut]"
      },
      "ResponseModel_AdminConfirmOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/AdminConfirmOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminConfirmOut]"
      },
      "ResponseModel_AdminRead_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/AdminRead"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminRead]"
      },
...
      "ResponseModel_RefreshSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/RefreshSessionOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[RefreshSessionOut]"
      },
...
      "ResponseModel_UserAuthSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/UserAuthSessionOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[UserAuthSessionOut]"
      },
...
      "ResponseModel_bool_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "type": "boolean",
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[bool]"
      },
...
      "RoleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "RoleSummary"
      },
...
      "UserAuthSessionOut": {
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/UserPrincipalOut"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "access_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Access Expires At"
          },
          "idle_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Absolute Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "principal",
          "session_id",
          "access_expires_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "UserAuthSessionOut"
      },
      "UserLoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLoginIn"
      },
      "UserPrincipalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "email",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserPrincipalOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}